move Attachment classes to Attachment
