+++ frontend/src/Activity/Garmin.svelte
@@ -10,10 +10,11 @@
     InputGroupText,
     Label,
   } from "@sveltestrap/sveltestrap";
-  import { attachments, checkStatus, handleError } from "../lib/store";
+  import { checkStatus, handleError } from "../lib/store";
   import TZPicker from "../Widgets/TZPicker.svelte";
   import ModalFooter from "../Widgets/ModalFooter.svelte";
   import { parts } from "../Part/part";
+  import { attachments } from "../Attachment/attachment";
 
   let timezone: string | undefined;
