+++ frontend/src/Attachment/AttachPart.svelte
@@ -2,7 +2,7 @@
   import { Modal, ModalHeader, ModalBody } from "@sveltestrap/sveltestrap";
   import ModalFooter from "../Widgets/ModalFooter.svelte";
   import { types } from "../lib/store";
-  import { AttEvent } from "../lib/types";
+  import { AttEvent } from "./attachment";
   import AttachForm from "./AttachForm.svelte";
   import { Part } from "../Part/part";
