+++ frontend/src/Part/PartHist.svelte
@@ -1,9 +1,10 @@
 <script lang="ts">
   import { filterValues, by } from "../lib/mapable";
-  import { attachments, types } from "../lib/store";
+  import { types } from "../lib/store";
   import Usage from "../Usage/Usage.svelte";
   import { link } from "svelte-spa-router";
   import { parts } from "./part";
+  import { attachments } from "../Attachment/attachment";
 
   export let id: number;
