+++ frontend/src/lib/store.ts
@@ -1,10 +1,11 @@
 import { writable, type Writable } from "svelte/store";
 import { mapable, mapObject } from "./mapable";
-import { Attachment, Type, type ActType, type User } from "./types";
+import { Type, type ActType, type User } from "./types";
 import { Service, services } from "../Service/service";
 import { activities, Activity } from "../Activity/activity";
 import { Usage } from "../Usage/usage";
 import { parts, type Part } from "../Part/part";
+import { Attachment, attachments } from "../Attachment/attachment";
 
 export function fmtDate(date: Date | undefined) {
   return date ? date.toLocaleDateString(navigator.language) : "never";
@@ -128,11 +129,6 @@ export let types: { [key: number]: Type };
 export let category: Writable<Type>;
 export const user = writable<User | undefined>(undefined);
 
-export const attachments = mapable(
-  "idx",
-  (a) => new Attachment(a),
-  (a) => a.isEmpty(),
-);
 export const usages = mapable("id", (u) => new Usage(u));
 
 export const state = writable({ show_all_spares: false });
