+++ frontend/src/lib/types.ts
@@ -4,49 +4,6 @@ import { type Map, by, filterValues } from "./mapable";
 import { fmtDate, handleError, myfetch, updateSummary } from "./store";
 export const maxDate = new Date("2999-12-31");
 
-export class Attachment {
-  part_id: number;
-  attached: Date;
-  gear: number;
-  hook: number;
-  detached: Date;
-  what: number;
-  name: string;
-  idx: string;
-  usage: string;
-  constructor(data: any) {
-    this.part_id = data.part_id;
-    this.attached = new Date(data.attached);
-    this.gear = data.gear;
-    this.hook = data.hook;
-    this.detached = new Date(data.detached);
-    this.what = data.what;
-    this.name = data.name;
-    this.idx = this.part_id + "/" + this.attached.getTime();
-    this.usage = data.usage;
-  }
-  fmtTime() {
-    let res = fmtDate(this.attached);
-    if (this.detached < maxDate) res = res + " - " + fmtDate(this.detached);
-    return res;
-  }
-  isAttached(time?: Date | string | number) {
-    if (!time) time = new Date();
-    else time = new Date(time);
-
-    return this.attached <= time && time < this.detached;
-  }
-  isEmpty() {
-    return this.attached.getTime() >= this.detached.getTime();
-  }
-  activities(acts: Map<Activity>) {
-    return filterValues(
-      acts,
-      (a) => a.gear == this.gear && this.isAttached(a.start),
-    );
-  }
-}
-
 export class Type {
   id: number;
   name: string;
@@ -92,37 +49,3 @@ export type ActType = {
   name: string;
   gear_type: number;
 };
-
-export class AttEvent {
-  part_id: number;
-  time: Date;
-  gear: number;
-  hook: number;
-  constructor(
-    part: number | undefined,
-    time: Date,
-    gear: number | undefined,
-    hook: number,
-  ) {
-    if (gear == undefined || part == undefined) {
-      console.error("part or gear not defined: ", part, gear);
-      throw "part or gear not defined";
-    }
-    this.part_id = part;
-    this.time = time;
-    this.gear = gear;
-    this.hook = hook;
-  }
-
-  async attach() {
-    return await myfetch("/part/attach", "POST", this)
-      .then(updateSummary)
-      .catch(handleError);
-  }
-
-  async detach() {
-    return await myfetch("/part/detach", "POST", this)
-      .then(updateSummary)
-      .catch(handleError);
-  }
-}
